Small restructuring of project
