+++ environment.yml
@@ -9,6 +9,7 @@ channels:
 dependencies:
   - black=23
   - dacite
+  - diffusers>=0.28
   - dvc>=3
   - flake8=6
   - h5py>=3.7
