+++ favicon_gen/diffusion/diffuser_model.py
