+++ favicon_gen/vae/autoencoder.py
