+++ tests/diffusion/test_denoising_diffusion.py
@@ -5,7 +5,7 @@ import random
 import torch
 
 from favicon_gen.data_loading import LargeLogoDataset, show_image_grid
-import favicon_gen.denoising_diffusion as ddi
+import favicon_gen.diffusion.denoising_diffusion as ddi
 
 
 @pytest.fixture()
